+++ src/main/java/com/darigan/lil/learningspring/data/RoomRepository.java
@@ -2,5 +2,5 @@ package com.darigan.lil.learningspring.data;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoomRepository extends CrudRepository<Room, Long> {
+public interface RoomRepository extends CrudRepository<Room, Integer> {
 }
